+++ src/commonMain/kotlin/com/jessecorbett/diskord/api/rest/client/GuildClient.kt
@@ -6,7 +6,7 @@ import com.jessecorbett.diskord.api.rest.*
 import com.jessecorbett.diskord.api.rest.client.internal.DefaultRateLimitedRestClient
 import com.jessecorbett.diskord.api.rest.client.internal.RateLimitedRestClient
 import com.jessecorbett.diskord.util.DiskordInternals
-import kotlinx.serialization.list
+import kotlinx.serialization.builtins.list
 
 /*
  * Note: Emoji don't follow standard rate limit behavior, and the API responses may not accurately reflect rate limits.
@@ -22,7 +22,7 @@ import kotlinx.serialization.list
  * @param guildId The id of the guild.
  * @param userType The user type, assumed to be a bot.
  */
-@UseExperimental(DiskordInternals::class)
+@OptIn(DiskordInternals::class)
 class GuildClient(
     token: String,
     val guildId: String,