+++ src/commonMain/kotlin/com/jessecorbett/diskord/dsl/Core.kt
@@ -28,7 +28,7 @@ annotation class DiskordDsl
  * @param shardId The shard id, if this bot is sharded.
  * @param shardCount The count of shards, if the bot is sharded.
  */
-@UseExperimental(KtorExperimentalAPI::class)
+@OptIn(KtorExperimentalAPI::class)
 class Bot(token: String, shardId: Int = 0, shardCount: Int = 0) : EnhancedEventListener(token) {
     private val websocket = DiscordWebSocket(token, this, shardId = shardId, shardCount = shardCount)